deepl/ai-training-platform/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
opencv-python
pyyaml
gputil
sqlalchemy
shapely
pyclipper
onnxruntime
pdf2image
pillow
torch==2.0.1
torchvision==0.15.2
ultralytics
sentencepiece
matplotlib
tqdm
psutil
psycopg2-binary==2.9.10
numpy==1.24.0
pyyaml==6.0.1
pillow==10.2.0
# onnxruntime-gpu==1.17.1
opencv-python-headless
pytesseract==0.3.10
python-dotenv==1.0.1
protobuf==5.26.1
onnx>=1.12.0
onnxslim
onnxruntime-gpu

# annotated-types==0.7.0
# anyio==4.3.0
# appdirs==1.4.4
# async-timeout==4.0.3
# certifi==2024.2.2
# charset-normalizer==3.3.2
# click==8.1.7
# coloredlogs==15.0.1
# dnspython==2.6.1
# easydict==1.13
# ecdsa==0.19.0
# exceptiongroup==1.2.1
# fastapi==0.111.0
# fastapi-cli==0.0.4
# filelock==3.13.1
# flatbuffers==24.3.25
# fsspec==2024.2.0
# h11==0.14.0
# httpcore==1.0.5
# httptools==0.6.1
# httpx==0.27.0
# humanfriendly==10.0
# idna==3.7
# imageio==2.34.1
# jinja2==3.1.3
# joblib==1.4.2
# lazy-loader==0.4
# mako==1.3.5
# markdown-it-py==3.0.0
# markupsafe==2.1.5
# mdurl==0.1.2
# mpmath==1.3.0

# onnx==1.16.0


# orjson==3.10.3
# packaging==24.0
# paho-mqtt==2.1.0
# passlib==1.7.4

# platformdirs==4.2.2

# pyasn1==0.6.0
# pycuda #==2024.1
# pydantic==2.7.1
# pydantic-core==2.18.2
# pygments==2.18.0


# python-jose==3.3.0
# python-multipart==0.0.9
# pytools==2024.1.3

# redis==5.0.4
# requests==2.31.0
# rich==13.7.1
# rsa==4.9
# scikit-image==0.22.0
# scikit-learn==1.4.2
# scipy
# sentencepiece==0.2.0
# shellingham==1.5.4
# six==1.16.0
# sniffio==1.3.1
# starlette==0.37.2
# sympy==1.12
# tensorrt==8.5.3.1
# threadpoolctl==3.5.0
# tifffile==2024.5.10

# torchaudio==2.0.2
# typer==0.12.3
# typing-extensions==4.9.0
# ujson==5.10.0
# urllib3==2.2.1
# uvicorn==0.29.0
# uvloop==0.19.0
# watchfiles==0.21.0
# websockets==12.0

